INV-141: refactor module start
