+++ server/app/src/main.rs
@@ -21,7 +21,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing_subscriber::fmt::init();
 
     let app = Router::new()
-        .route("/", get(root))
+        .route("/", get(health))
         .route("/items", get(get_items))
         .route("/places", get(get_places))
         .route("/inventory/items", get(get_inventory_items))
@@ -35,302 +35,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     Ok(())
 }
 
-async fn root() -> &'static str {
-    "Hello, World!"
-}
-
-async fn get_items_db(pool: &PgPool) -> Result<Vec<Item>, Box<dyn std::error::Error>> {
-    let items = sqlx::query!("SELECT item_id, item_name FROM Items ORDER BY item_name;")
-        .fetch_all(pool)
-        .await?
-        .into_iter()
-        .map(|record| Item {
-            item_id: record.item_id,
-            item_name: record.item_name,
-        })
-        .collect();
-
-    Ok(items)
-}
-
-async fn get_places_db(pool: &PgPool) -> Result<Vec<Place>, Box<dyn std::error::Error>> {
-    let places =
-        sqlx::query!("SELECT place_id, place_name, place_type FROM Places ORDER BY place_name;")
-            .fetch_all(pool)
-            .await?
-            .into_iter()
-            .map(|record| Place {
-                place_id: record.place_id,
-                place_name: record.place_name,
-                place_type: record.place_type,
-            })
-            .collect();
-
-    Ok(places)
-}
-
-async fn get_items(Extension(pool): Extension<PgPool>) -> impl IntoResponse {
-    match get_items_db(&pool).await {
-        Ok(items_result) => {
-            let list_response: Vec<Value> = items_result
-                .into_iter()
-                .map(|item| {
-                    let mut map = Map::new();
-                    map.insert(
-                        "item_id".to_string(),
-                        Value::String(item.item_id.to_string()),
-                    );
-                    map.insert("item_name".to_string(), Value::String(item.item_name));
-                    Value::Object(map)
-                })
-                .collect();
-            let obj = Value::Array(list_response);
-            (StatusCode::OK, Json(obj))
-        }
-        Err(err) => {
-            eprintln!("Error fetching items: {}", err);
-            let error_response = serde_json::json!({ "error": err.to_string() });
-            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
-        }
-    }
-}
-
-async fn get_places(Extension(pool): Extension<PgPool>) -> impl IntoResponse {
-    match get_places_db(&pool).await {
-        Ok(places_result) => {
-            let list_response: Vec<Value> = places_result
-                .into_iter()
-                .map(|place| {
-                    let mut map = Map::new();
-                    map.insert(
-                        "place_id".to_string(),
-                        Value::Number(Number::from(place.place_id)),
-                    );
-                    map.insert("place_name".to_string(), Value::String(place.place_name));
-                    map.insert("place_type".to_string(), Value::String(place.place_type));
-                    Value::Object(map)
-                })
-                .collect();
-            let obj = Value::Array(list_response);
-            (StatusCode::OK, Json(obj))
-        }
-        Err(err) => {
-            eprintln!("Error fetching places: {}", err);
-            let error_response = serde_json::json!({ "error": err.to_string() });
-            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
-        }
-    }
-}
-
-async fn get_inventory_items(
-    Extension(pool): Extension<PgPool>,
-    query: Query<InventoryItemQuery>,
-) -> impl IntoResponse {
-    match get_inventory_items_db(&pool, &query).await {
-        Ok(items_result) => {
-            let list_response: Vec<Value> = items_result
-                .into_iter()
-                .map(|item| {
-                    let mut map = Map::new();
-                    map.insert(
-                        "item_id".to_string(),
-                        Value::Number(Number::from(item.item_id)),
-                    );
-                    map.insert("item_name".to_string(), Value::String(item.item_name));
-                    map.insert(
-                        "nb_of_items".to_string(),
-                        Value::Number(Number::from(item.nb_of_items)),
-                    );
-                    Value::Object(map)
-                })
-                .collect();
-            let obj = Value::Array(list_response);
-            (StatusCode::OK, Json(obj))
-        }
-        Err(err) => {
-            eprintln!("Error fetching inventory items: {}", err);
-            let error_response = serde_json::json!({ "error": err.to_string() });
-            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
-        }
-    }
-}
-
-async fn get_inventory_places(
-    Extension(pool): Extension<PgPool>,
-    query: Query<InventoryPlaceQuery>,
-) -> impl IntoResponse {
-    match get_inventory_places_db(&pool, &query).await {
-        Ok(places_result) => {
-            let list_response: Vec<Value> = places_result
-                .into_iter()
-                .map(|place| {
-                    let mut map = Map::new();
-                    map.insert(
-                        "place_id".to_string(),
-                        Value::Number(Number::from(place.place_id)),
-                    );
-
-                    map.insert("place_name".to_string(), Value::String(place.place_name));
-                    map.insert("place_type".to_string(), Value::String(place.place_type));
-                    map.insert(
-                        "nb_of_items".to_string(),
-                        Value::Number(Number::from(place.nb_of_items)),
-                    );
-                    Value::Object(map)
-                })
-                .collect();
-            let obj = Value::Array(list_response);
-            (StatusCode::OK, Json(obj))
-        }
-        Err(err) => {
-            eprintln!("Error fetching inventory places: {}", err);
-            let error_response = serde_json::json!({ "error": err.to_string() });
-            (StatusCode::INTERNAL_SERVER_ERROR, Json(error_response))
-        }
-    }
-}
-
-async fn get_inventory_items_db(
-    pool: &PgPool,
-    query: &Query<InventoryItemQuery>,
-) -> Result<Vec<InventoryItem>, Box<dyn std::error::Error>> {
-    let default = "";
-    let is_query_empty = query.place_name.as_deref().unwrap_or(default) == default
-        && query.place_type.as_deref().unwrap_or(default) == default;
-    let items = match is_query_empty {
-        true => sqlx::query!(
-            "SELECT Items.item_id, Items.item_name, SUM (Inventory.nb_of_items) AS nb_of_items
-                    FROM Inventory
-                    JOIN Places ON Inventory.place_id = Places.place_id
-                    JOIN Items ON Inventory.item_id = Items.item_id
-                GROUP BY Items.item_id, Items.item_name 
-                ORDER BY nb_of_items DESC, Items.item_name;"
-        )
-        .fetch_all(pool)
-        .await?
-        .into_iter()
-        .map(|record| InventoryItem {
-            item_id: record.item_id,
-            item_name: record.item_name,
-            nb_of_items: record.nb_of_items.unwrap_or(0) as i32,
-        })
-        .collect(),
-        _ => sqlx::query!(
-            "SELECT Items.item_id, Items.item_name, Inventory.nb_of_items
-                    FROM Inventory
-                    JOIN Places ON Inventory.place_id = Places.place_id
-                    JOIN Items ON Inventory.item_id = Items.item_id
-                    WHERE (place_name =  $1 OR $1 = '') 
-                        AND (place_type = $2 OR $2 = '')
-                    ORDER BY Inventory.nb_of_items DESC;",
-            query.place_name.as_deref().unwrap_or(default),
-            query.place_type.as_deref().unwrap_or(default),
-        )
-        .fetch_all(pool)
-        .await?
-        .into_iter()
-        .map(|record| InventoryItem {
-            item_id: record.item_id,
-            item_name: record.item_name,
-            nb_of_items: record.nb_of_items,
-        })
-        .collect(),
-    };
-
-    Ok(items)
-}
-
-async fn get_inventory_places_db(
-    pool: &PgPool,
-    query: &Query<InventoryPlaceQuery>,
-) -> Result<Vec<InventoryPlace>, Box<dyn std::error::Error>> {
-    let default = "";
-    let is_query_empty = query.item_name.as_deref().unwrap_or(default) == default;
-    let places = match is_query_empty {
-        true => {
-            sqlx::query!(
-                "SELECT Places.place_id, Places.place_name, Places.place_type, SUM(Inventory.nb_of_items) as nb_of_items
-                    FROM Inventory
-                    JOIN Places ON Inventory.place_id = Places.place_id
-                    JOIN Items ON Inventory.item_id = Items.item_id
-                GROUP BY Places.place_id, Places.place_name, Places.place_type
-                ORDER BY nb_of_items DESC, Places.place_name;"
-            )
-            .fetch_all(pool)
-            .await?
-            .into_iter()
-            .map(|record| InventoryPlace {
-                place_id: record.place_id,
-                place_name: record.place_name,
-                place_type: record.place_type,
-                nb_of_items: record.nb_of_items.unwrap_or(0) as i32,
-            })
-            .collect()
-        },
-        _ => {
-            sqlx::query!(
-                "SELECT Places.place_id as place_id, Places.place_name as place_name, Places.place_type as place_type, Inventory.nb_of_items as nb_of_items
-                    FROM Inventory
-                    JOIN Places ON Inventory.place_id = Places.place_id
-                    JOIN Items ON Inventory.item_id = Items.item_id
-                    WHERE (item_name =  $1 OR $1 = '') 
-                    ORDER BY Inventory.nb_of_items DESC;",
-                query.item_name.as_deref().unwrap_or(default),
-            )
-            .fetch_all(pool)
-            .await?
-            .into_iter()
-            .map(|record| InventoryPlace {
-                place_id: record.place_id,
-                place_name: record.place_name,
-                place_type: record.place_type,
-                nb_of_items: record.nb_of_items,
-            })
-            .collect()
-        }
-    };
-
-    Ok(places)
-}
-
-#[derive(Serialize)]
-struct Item {
-    item_id: i32,
-    item_name: String,
-}
-
-#[derive(Serialize)]
-struct Place {
-    place_id: i32,
-    place_name: String,
-    place_type: String,
-}
-
-#[derive(Serialize)]
-struct InventoryItem {
-    item_id: i32,
-    item_name: String,
-    nb_of_items: i32,
-}
-
-#[derive(Serialize, Debug)]
-struct InventoryPlace {
-    place_id: i32,
-    place_name: String,
-    place_type: String,
-    nb_of_items: i32,
-}
-
-#[derive(Deserialize, Debug)]
-struct InventoryPlaceQuery {
-    item_name: Option<String>,
-}
-#[derive(Deserialize, Debug)]
-struct InventoryItemQuery {
-    place_name: Option<String>,
-    place_type: Option<String>,
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -340,7 +44,7 @@ mod tests {
 
     #[tokio::test]
     async fn test_root() {
-        let app = Router::new().route("/", get(root));
+        let app = Router::new().route("/", get(health));
 
         let response = app.oneshot(Request::new(Body::empty())).await.unwrap();
