+++ server/resource/Cargo.toml
@@ -4,3 +4,10 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+repository = { path = "../repository/src/lib.rs" }
+
+sqlx = { version = "0.7.4", features = ["runtime-tokio", "postgres"] }
+tokio = { version = "1", features = ["full"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+axum = "0.7.5"
